requirements.txt.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
django~=4.1.7
django-allauth==0.52.0
django-countries==7.5.1
django-crispy-forms==1.9.2
django-debug-toolbar==3.8.1
idna==2.8
oauthlib==3.0.1
pep8==1.7.1
Pillow==9.4.0
pycodestyle==2.5.0
python-decouple==3.1
python3-openid==3.1.0
requests==2.21.0
gunicorn
pytz==2020.1
django-widget-tweaks==1.4.8